requirements.txt
PyYAML
schedule
discord-webhook